Backend/MoodModel/requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
torchvision
Pillow
